+++ tests/test_integration/test_client/test_utils/test_dataset_calls.py
@@ -4,7 +4,7 @@ from cognite.client.data_classes import DataSet
 from cognite.powerops.client._generated.v1.data_classes import DataSetConfiguration
 from cognite.powerops.client._generated.v1.data_classes._data_set_configuration import DataSetConfigurationWrite
 from cognite.powerops.client.powerops_client import PowerOpsClient
-from cognite.powerops.utils.cdf.datasets_calls import get_data_set_from_config
+from cognite.powerops.utils.retrieve import get_data_set_from_config
 from tests.test_integration.conftest import random_external_id